tg-order-bot/requirements.txt
# 依赖包
python-telegram-bot>=20.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pillow>=10.0.0
